Add files via upload
Using Python, TensorFlow. we built a deep learning model that analyzes chest X-ray images to detect pneumonia. The data was preprocessed and augmented for better performance. A custom CNN was trained on labeled X-ray images, classifying them as either "Pneumonia" or "Normal". After training, the model was evaluated using confusion matrices and classification reports. The final model can accurately predict the presence of pneumonia in real-time, and can be extended for deployment in healthcare diagnostics.